requirements.txt
chainlit>=0.7.700
groq>=0.4.6
Pillow>=10.0.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
asyncio>=3.4.3
typing-extensions>=4.8.0
numpy>=1.24.0